Change buckets/indices for core store.
Descriptions were stored twice - firstly in a 'descriptions' bucket keyed by descriptionId, and secondly in a 'concept-descriptions' bucket, keyed by a tuple ('conceptId'-'descriptionId'). This increases memory usage unnecessarily and paradoxically may be slower given the mmap characteristics of mapdb in general usage. So, this change optimises for the common usage - get descriptions for a concept, and uses a simple ('descriptionId'-'conceptId') index for descriptionId lookup, which has a much less frequent use case, if one at all in normal usage.