Add quick web service and CLI interfaces.
Add a simple bare-bones pedestal based REST server and similarly bare-bones CLI to start it up. Create a unified terminology abstraction that wraps the 'store' and 'search' functionality behind a single API. Add manifest so we can track versioning and emit an error early if formats of either store or search indices are incorrect.