Fix handling of nested exclusion constraints.
When processing a flat expression, we can directly apply exclusions. If that expression is nested, say in an attribute (property=exp), then we need to decompose the expression and re-write as (and (property=included) (property!=excluded). So we take the expression, split into 'included' and 'excluded' components, and re-combine at the higher level of context.