First hack at dm+d extension
The SNOMED dm+d release doesn't include everything that is in the canonical dm+d XML release files, such as the numeric doses of ingredients in a given product. But SNOMED provides an ontology around entities such as substances. Blending the two with an extension makes sense, at least until all dm+d data is provided in SNOMED release format (perhaps lack of concrete types was a limiting factor). So plan to build an extension mechanism to add namespaced extension data to any concept, and informed clients can then make the request for data in an extension (e.g. give me the dm+d version of this VMP... and it will include the dm+d data as well as all of the SNOMED data.