Add grouping to relationship logic.
For pure ontological reasoning, we can simply use all parent relationships and not be concerned about how different relationships relate to one another. However, for other uses, we need to collect relationships together when they are related - for example, in the UK dm+d, we'll have groups of relationships for different ingredients in a compound product. So, this tweaks index creation to include the grouping, and then adds a simple way of returning those grouped properties.