Add query re-writing when single 'must-not' clause used.
Lucene can't return any results if there is a single 'must-not' clause. This doesn't usually cause any great difficulties, as usually there is always a MUST or SHOULD clause as well, such as including only active concepts. But in the rare case of wanting to search inactive concepts, we need to rewrite a query to ensure a single 'must-not' clause is either flattened into the parent query, or combined with a 'match-all-docs'. When flattened into the parent query, this will still return no results if there is a single 'must-not' clause - unless the client specifically chooses to match all documents - this is by design as otherwise pretty much the whole index will be returned!