Add support for cardinality range including zero.
This adds support for searching for an attribute with a range of [0..x]. The number of attributes for any given concept is stored as a field "cXXXXXX" where XXXXXX is the attribute concept identifier. It is therefore straightforward to find a range [1...x] - but we have to negate using match-all-docs in order to include 0 in our range. This is potentially a problem for performance, which could be avoided by increasing index size and recording a count of zero for all known attributes at time of index creation. If a compelling use case becomes evident, this could be a reasonable strategy. For now, this is a reasonable implementation that is performant when combined with other queries.