Actually commit fix to SNOMED filename parsing.
Fix change in UK release filename specification to cope with additional random information in the summary field of the content-subtype field. Switch to using named capture groups to make it easier to adapt in the future, and it is probably more robust now to future changes, although unfortunately it means reference set types are now hard-coded as NHS Digital's change permits no other approach.