Conveniently fallback to using en-US if no match for locale
Hermes tries to make it easy by building a cache of preferred descriptions for search and autocompletion. This isn't necessary, but helpful. It usually uses the default locale to choose, but we can now define the locale to use at time of index creation. However, if we do not have a map from locale to a set of language reference sets for a specific locale, this now falls back to using en-US rather than failing with what might appear to be a weird exception.