Better handling of errors during import
Fixes #28. If we throw an exception in a thread, it was not caught and the thread would silently end. This creates an error channel onto which exceptions can be forwarded back to the main thread, aborting the import process. The exit status of hermes will be non-zero, and so will properly abort any script. It's conceivable that we could add a --keep-going flag which simply logs those errors and tries to continue.