Add new extended refset, for custom refsets.
This can, if given a pattern, store arbitrary data over and above the core simple reference set. Fix identification of the type of refset. Instead of relying on the name of refset type being in the summary, we now assume a simple refset, or an extended refset if there is a pattern prefix. This fixes #31 and partially helps #30, although relies on distributions having the right naming prefix.