Add graph API tests
These demonstrate the declarative power of the the pathom graph API. We specify the data we want, and the data are returned in the 'shape' as requested. A significant benefit is decoupling, so that other resolvers can be composed with these in order to extend functionality far beyond what is provided in this library. For example, walking from a patient record, through SNOMED CT to the UK dm+d data through deprivation data. Decoupling but seamless composition are key.