Fix handling of extended reference items
Hermes implements reference set items as concrete subtypes, because it makes use of those data in its APIs. The other option would have been a completely generic reference set item, that could take any number of fields with properties determined during import or at runtime using the metadata model. In addition, any local distribution can provide additional fields to extend a core reference set type.

This commit adds a :fields property to all reference set types. This means all are extendable at runtime, while keeping the core information for that type available for processing easily as part of the APIs provided.

This means it will be easy to make these extended properties available at runtime, with only a small overhead to check the metadata model for that reference set.

The alternatives would have been to:

a) make all reference set items generic and have dynamic properties outside of the core 6 properties.

This would make serialisation a little more complex [although not onerously so] and make using those data items slightly more difficult [again, not onerously given the dynamic nature of clojure, but more difficult for clients using more statically typed languages such as java] and force serialisation of not only values but also keys for all refset properties outside of the core, increasing database size.

b) serialise key value pairs for each extended property outside of the core for that subtype.

This would be subject to increasing serialization costs (time and size).

As such, the compromise is to keep concrete versions of each reference set subtype, and simply store a vector of the extended properties, and we will use the metadata model to turn these into key value pairs on demand. This is flexible, serialising according to the refset 'pattern' and yet future proofing for any changes.