Add runtime refset item reification
This is needed for a very small number of refsets that do not include the concrete refset type in their filename and are therefore not reified at time of import. At the moment, import does not try to do reification unless the concrete type is in the filename, although that could change in the future if import used a two-pass approach so that refset descriptor data were available at the time of import. Instead, this adds an ability to transparently reify a sequence of refset items using runtime information from the refset descriptors.