Extract out mapdb dependent code
In preparation for evaluating different backend key value stores, extract out core 'key value store' functionality without creating an unnecessary layer of abstraction. While the latter is possible, with 'buckets' and 'indexes', this is performance-critical code, and it is clear from testing that it is sensible to take advantage of the features of specific stores.