Remove deprecated flag and Improve docstring for reverse map fns
While these are not necessarily recommended, they don't have to be deprecated formally. Instead, simply improve docstring to recommend a better alternative for most use cases. Also, because these map from a target code, to a number of concepts, and from there, to a number a target codes, you end up with non-intuitive results in which the resulting target does not meet the original criteria. As such, these must subsequently re-apply the original criteria. The real solution to this is to use member-field-prefix (or member-field, or member-field-wildcard) and then determine what to do with those results on a case-by-case basis, as they only do the target code->concepts step, rather than report using refset items.