Add new command-line parsing module
This is backwards-compatible with prior operation, but provides greater flexibility in specifying options depending on command and other options, dynamically. In addition, these functions are now 'pure', taking in data and returning data, making it easier to test, and easier to use. Only options that are appropriate for the context are now shown, which makes it possible to add more specific options to certain commands while not overcomplicating basic usage. While command 'download' is maintained for backwards compatibility, it is deprecated in favour of 'install'.