Better defaults for environment map sizes
In typical usage, this means we have core and refsets environments at <70% of capacity. The general advice is to set environment map size well in excess of likely requirements, and spare files are used by default, but there is little point in grossly overestimating. In read-only mode, avoid setting a map size manually, so the memory mapped file will be the same size as the data.