Speed up transitive parent relationships, and therefore index creation
Benchmarking showed that building the search index, and generating denormalised 'extended' concepts for in the core library API, and the HTTP API, depended heavily on 'get-parent-relationships-expanded' that looks at a concept's parent relationships and generates a transitive closure table for each. The prior implementation was pretty naive, so this change speeds it up by 1.8x (176ms vs 313ms) and overall description search index creation by 30 seconds.