Avoid deprecated methods when using Lucene >=9.5
To maintain Java 1.8 compatibility, we offer the option to use Lucene 8. However, there are some changes in Lucene 9.5 such as the recommended way to access storedFields. These new APIs apparently perform better under heavy load and concurrency, so results will be faster and have less memory overhead. It is also very likely that deprecated methods will be removed in future versions, so this takes advantage of the dynamic nature of Clojure to fallback to a different implementation depending on Lucene version. I am not clear whether the complexity this introduces is worth it, and given the nature of Hermes, it should simply track modern Java and latest versions of dependencies. See #34 and #43.