Handle special case of cardinality [0..0] and wildcard values
When handling refinements, we can transform wildcard values to a count of attributes query, such that a search is performed for results with >0 attributes matching the query. However, when used in conjunction with a [0..0] cardinality, we must reverse this wildcard filter so that we actually exclude results with >0 attributes. When the expression contains a minimum cardinality of zero, we are careful to avoid adding an attribute count query, and for each value, simply exclude results with an attribute count greater than the maximum requested. Fixes #46.