Guard against open ended cardinality constraint
While the ECL specifications state: "A cardinality of [0..*] should therefore never be used as this indicates that the given attribute is not being constrained in any way, and is therefore a redundant part of the expression constraint.", we must handle this case and guard against an integer overflow. Alternative of throwing an exception during ECL parsing, because this is an invalid cardinality considered but rejected as an option.