Add support for concrete values to key value store
Although concrete values are modelled after the Relationship entity, they are semantically quite different. As such, rather than keeping concrete values in the same DBI as all of the Relationship entities, and needing separate indexes, the common usage for concrete values is to lookup using a concept identifier, much like we do for descriptions. Therefore, we optimise for that usage pattern, with a new DBI keyed by a composite key of concept-id--relationship-id, permitting fast lookup of all of the concrete values for a given concept. All values are very likely to be in the same mmap'ed memory page given how LMDB works.

As this is a work-in-progress, the backing store treats the new concrete values DBI as optional, flagging a warning only for diagnostic purposes if a legacy database is used.