Add MRCM-enabled 'properties' to top level API
This takes the combination of relationships and concrete values, and information from the MRCM (machine readable concept model) relating to appropriate cardinality, and builds a property map with grouped attributes and values, to make it as easy as possible to consume for clients. These additional functions are marked private for the time being. To optimise access to the concept model, we add an appropriate closure into the main service (gives 10x speedup; 40ms->4ms in benchmarking). Depending on need, additional data from the MRCM could also be pre-computed like this.