Improve HTTP server
Avoid HTTP 500 errors if an interceptor returns nil. Instead, return a 404 by returning an unchanged context, and thus avoid confusing users / clients. For endpoints returning collections (e.g. get-concept-descriptions), take care to return 404, an empty collection or a collection containing items depending on parameters. For example, it is incorrect to return an empty collection for a concept that does not exist.