Update `properties` to optionally include transitive relationships
This now includes all properties, including concrete values, in groups. Optionally, values will be expanded to include transitive relationships. If there is no expansion, or the value is a concrete value, and the metadata model states that in the context the property has a cardinality of one, a set containing only one value will be unwrapped to a single value.