Throw for memberOf when used to select other fields
This is currently unimplemented, so rather than failing silently, throw an exception. It is fairly easy to implement this, but the sub expression constraint will have to deal with this special case in which results are no longer a query that can be used against the search index, but a query for reference set items or the expanded result of those items. It may be that for consistency, all parsing should return a map of data rather than only a query, if this is considered important enough to implement.