Use macro to do more work at compile-time
This provides a 12.5% speed-up for, e.g. 'all-parents' as a consequence of moving some logic to compile time from runtime. This now avoids unnecessary marshalling of data into and out of a vector, and instead writes those data directly into a ByteBuf for using a cursor to iterate over a range of keys in a backing DBI.