Test for data integrity issues during indexing
As per #64, indexing currently uses two sequential write transactions which delete the
relationship and refset indices, and then rebuild those indices. That means if there are
concurrent threads or processes may read inconsistent data during an update-in-place operation.
This tests for this issue and will result in a failing test if data are inconsistent during an
indexing operation.