Better documentation for store/leaves and add store/top-leaves
When the SNOMED CT hierarchy is drawn top down with the root at the top,
as is usual in computer science, higher up concepts are more general,
and lower down concepts are more specific. SNOMED documentation
introduces 'top-of-set' and 'bottom-of-set', but we already have
'leaves' which provides the latter. As such, this commit improves the
documentation and adds 'top-leaves' for 'top-of-set'. Performance
testing confirms that this is 'top-of-set' is however very slow, as
expected. It is better to plan to use the search index to implement the
real 'top-of-set' ECL operator, but this will be helpful in validating
results during testing. #67