Migrate to Lucene 10
Lucene 10 requires Java 21 or higher, which is entirely reasonable, but Hermes is currently used in a variety of environments in which this might be a difficult pre-requisite to reach at the moment. There are few breaking changes but one that affects Hermes is this change in BooleanClause from a class to a record type, which means field access changes.