Add option to run using Lucene 10 for testing
Lucene 10 requires Java 21, so we need to support Lucene 9 for a little
while yet. The only incompatibility is in making BooleanClause a record
class, and no deprecated backwards compatible accessors have been
provided, which is frustrating. This therefore conditionally compiles
depending on the current Lucene version. This is an alternative to
maintaining a separate Lucene 10 branch.